requirements.txt
streamlit
requests
python-dotenv
langchain
langgraph

# Optional
# faiss-cpu
# scikit-learn
